+++ backend/src/main/java/com/storerader/server/common/repository/StoreRepository.java
@@ -8,5 +8,5 @@ import java.util.List;
 
 public interface StoreRepository extends JpaRepository<StoreEntity, Long> {
     @Query("select s.storeId from StoreEntity s")
-    List<Integer> findAllStoreIds();
+    List<Long> findAllStoreIds();
 }
